requirements.txt
jupyterlab
opencv-python
sk-video
webdataset
pytorch-lightning==1.7
einops
ray[tune]
ray[default]
optuna 
scenedetect[opencv]
ipywidgets
torchvision
ffmpeg-python
loguru
transformers
decord
moviepy